+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Unreleased
+
+* **terra-filter** is now deprecated . Use `<tc-filter>` instead.
+
 <a name="3.0.0></a>
 # 3.0.0 (18.12.2018)